server/requirements.txt
numpy==1.14.*
opencv-python==3.4.*
Pillow==5.1.*
sanic==0.7.*
scikit-learn==0.18.*
scipy==1.0.*
